requirements.txt
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
swagger-spec-validator==2.7.3
whitenoise==5.1.0
